requirements.txt
langchain==0.2.11
langchain-community==0.2.10
langchain-chroma==0.1.2
pypdf==4.3.1
docx2txt==0.8
ollama==0.3.0
python-dotenv==1.0.1
colorama==0.4.6
chromadb==0.5.5
tqdm==4.66.4